+++ src/signOpCert.h
@@ -3,6 +3,7 @@
 
 #ifdef APP_FEATURE_OPCERT
 
+#include "cardano.h"
 #include "common.h"
 #include "handlers.h"
 #include "bip44.h"
@@ -18,7 +19,7 @@ typedef struct {
 	uint64_t kesPeriod;
 	uint64_t issueCounter;
 	bip44_path_t poolColdKeyPathSpec;
-	uint8_t signature[64];
+	uint8_t signature[ED25519_SIGNATURE_LENGTH];
 	int ui_step;
 } ins_sign_op_cert_context_t;